% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{msg_done}
\alias{msg_done}
\alias{msg_todo}
\alias{msg_info}
\alias{msg_warn}
\alias{msg_oops}
\alias{msg_line}
\alias{msg_rule}
\alias{msg_field}
\alias{msg_value}
\alias{msg_code}
\title{Pre-formatted messages for a custom user interface}
\usage{
msg_done(...)

msg_todo(...)

msg_info(...)

msg_warn(...)

msg_oops(...)

msg_line(...)

msg_rule(...)

msg_field(...)

msg_value(...)

msg_code(...)
}
\arguments{
\item{...}{One or several expression (character or vector object).}
}
\value{
Print a success message.
}
\description{
These functions are used to construct an user interface \emph{a la} \code{usethis}.

These \verb{msg_*} functions can be broken down into two main categories:
\itemize{
\item \strong{block styles}: \code{msg_line()}, \code{msg_rule()}, \code{msg_done()}, \code{msg_todo()},
\code{msg_info()}, \code{msg_warn()}, \code{msg_oops()}.
\item \strong{inline styles}: \code{msg_field()}, \code{msg_value()}, \code{msg_code()}.
}
}
\section{Functions}{
\itemize{
\item \code{msg_todo}: Print a to do message.

\item \code{msg_info}: Print an informational message.

\item \code{msg_warn}: Print a warning message.

\item \code{msg_oops}: Print an error message.

\item \code{msg_line}: Print an (non-)empty message.

\item \code{msg_rule}: Print a rule message.

\item \code{msg_field}: Returns a customized code expression.

\item \code{msg_value}: Returns a customized variable value.

\item \code{msg_code}: Returns a customized code expression.
}}

\note{
Messages can be suppressed by using \code{options(messages = FALSE)}.
Messages can be reactivated by using: \code{options(messages = TRUE)}.
}
\examples{
## Block messages ----

msg_done("A success message")
msg_todo("A to do message")
msg_info("An informational message")
msg_warn("A warning message")
msg_oops("An error message")

msg_line()
msg_line("A message")

msg_rule("Left message")
msg_rule("Left message", , "Right message")
msg_rule("Left message", right = "Right message")
msg_rule(, "Center message")
msg_rule(center = "Center message")

## Inline messages ----

msg_done("The variable", msg_field("x"), "has been set to", msg_value(1))
msg_todo("Please use the function:", msg_code("msg_rule()"))
msg_line("Thanks for using", msg_code("messages"))
msg_line(msg_code("msg_rule()"))
}
